+++ bin_set.c
@@ -1,6 +1,6 @@
 #include "bin.h"
 
-void bin_set(char *vc, unsigned int n, int c)
+void bin_set(char *vc, unsigned long n, int c)
 {
   for (;;) {
     if (!n) return; *vc++ = c; --n;
