+++ bin_zero.c
@@ -1,6 +1,6 @@
 #include "bin.h"
 
-void bin_zero(char *vc, unsigned int n)
+void bin_zero(char *vc, unsigned long n)
 {
   for (;;) {
     if (!n) return; *vc++ = 0; --n;
