+++ buffer.h
@@ -6,25 +6,25 @@
 
 typedef struct {
   char *buf;
-  unsigned int pos;
-  unsigned int size;
+  unsigned long pos;
+  unsigned long size;
   int fd;
-  int (*op)(int, char *, unsigned int);
+  int (*op)(int, char *, unsigned long);
 } buffer;
 
 #define buffer_INIT(op,fd,buf,len) { (buf), 0, (len), (fd), (op) }
 
-void buffer_init(buffer *, int (*)(), int, char *, unsigned int);
+void buffer_init(buffer *, int (*)(), int, char *, unsigned long);
 
-int buffer_get(buffer *, char *, unsigned int);
+int buffer_get(buffer *, char *, unsigned long);
 int buffer_feed(buffer *);
 char *buffer_peek(buffer *);
-void  buffer_seek(buffer *, unsigned int);
+void  buffer_seek(buffer *, unsigned long);
 
 int buffer_flush(buffer *);
-int buffer_put(buffer *, const char *, unsigned int);
+int buffer_put(buffer *, const char *, unsigned long);
 int buffer_puts(buffer *, const char *);
-int buffer_putflush(buffer *, const char *, unsigned int);
+int buffer_putflush(buffer *, const char *, unsigned long);
 int buffer_putsflush(buffer *, const char *);
 
 int buffer_copy(buffer *, buffer *);
