+++ buffer_init.c
@@ -1,7 +1,7 @@
 #include "buffer.h"
 
-void buffer_init(buffer *b, int (*op)(int, char *, unsigned int),
-                 int fd, char *s, unsigned int len)
+void buffer_init(buffer *b, int (*op)(int, char *, unsigned long),
+                 int fd, char *s, unsigned long len)
 {
   b->size = len;
   b->pos = 0;
