+++ scan_newline.c
@@ -1,6 +1,6 @@
 #include "scan.h"
 
-unsigned int scan_newline(const char *s)
+unsigned long scan_newline(const char *s)
 {
   return scan_charset(s, "\n");
 }
